requirements.txt
torch
torchvision
opencv-python
pillow
matplotlib
numpy
pandas
scikit-learn
gradio              # ← ADD THIS (for Hugging Face)
jupyter
